feat: Added Slack mrkdown Nunjucks template to support Slack message format
