+++ src/az-servicehealth-integration/package.json
@@ -7,7 +7,7 @@
     "node": ">=18.0"
   },
   "scripts": {
-    "//": " && npm run cp:testdata", 
+    "//": " && npm run cp:testdata",
     "build": "tsc && npm run cp:nunjuckstpl && npm run mkdir:testdata",
     "watch": "tsc -w",
     "clean": "rimraf dist",
@@ -29,10 +29,12 @@
     "@azure/identity": "^4.4.0",
     "@azure/storage-queue": "^12.23.0",
     "@opentelemetry/api": "^1.9.0",
+    "@types/html-to-text": "^9.0.4",
     "@types/lodash": "^4.17.1",
     "@types/nodemailer": "^6.4.15",
     "@types/nunjucks": "^3.2.6",
     "dotenv": "^16.4.5",
+    "html-to-text": "^9.0.5",
     "lodash": "^4.17.21",
     "nodemailer": "^6.9.13",
     "nunjucks": "^3.2.4",
@@ -48,8 +50,8 @@
     "jest": "^29.7.0",
     "jest-environment-jsdom": "^29.7.0",
     "rimraf": "^5.0.0",
-    "ts-mockito": "^2.6.1",
     "ts-jest": "^29.2.5",
+    "ts-mockito": "^2.6.1",
     "tsx": "^4.10.2",
     "typescript": "^5.4.5"
   }
